requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==4.0.1
billiard==4.2.4
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==47.0.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.1
fastapi-cache2==0.2.2
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.13
iniconfig==2.3.0
Jinja2==3.1.6
kombu==5.6.2
Mako==1.3.11
MarkupSafe==3.0.3
packaging==26.2
passlib==1.7.4
pendulum==3.2.0
pillow==12.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.9.0
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
redis==7.4.0
rsa==4.9.1
ruff==0.15.12
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uvicorn==0.46.0
vine==5.1.0
wcwidth==0.6.0
